+++ server/domain/models/circle-invite-link/circle-invite-link-repository.ts
@@ -1,5 +1,5 @@
 import type { CircleId, InviteLinkToken } from "@/server/domain/common/ids";
-import type { CircleInviteLink } from "@/server/domain/models/circle/circle-invite-link";
+import type { CircleInviteLink } from "@/server/domain/models/circle-invite-link/circle-invite-link";
 
 export type CircleInviteLinkRepository = {
   findByToken(token: InviteLinkToken): Promise<CircleInviteLink | null>;
