+++ server/presentation/trpc/context.ts
@@ -11,7 +11,7 @@ import { prismaCircleSessionRepository } from "@/server/infrastructure/repositor
 import { prismaMatchRepository } from "@/server/infrastructure/repository/match/prisma-match-repository";
 import { prismaUserRepository } from "@/server/infrastructure/repository/user/prisma-user-repository";
 import { prismaSignupRepository } from "@/server/infrastructure/repository/user/prisma-signup-repository";
-import { prismaCircleInviteLinkRepository } from "@/server/infrastructure/repository/circle/prisma-circle-invite-link-repository";
+import { prismaCircleInviteLinkRepository } from "@/server/infrastructure/repository/circle-invite-link/prisma-circle-invite-link-repository";
 import { prismaUnitOfWork } from "@/server/infrastructure/transaction/prisma-unit-of-work";
 import {
   hashPassword,
